+++ test/unit/utils/helpers/case-body-helper-test.js
@@ -1,4 +1,4 @@
-import createBackOfficeCaseBody from "../../../lib/helpers/case-body-helper.js";
+import createBackOfficeCaseBody from "../../../../lib/helpers/case-body-helper.js";
 
 const description = "An order is in an unresolveable state for the system",
   namespace = "dn",
